packages/plugin-livekit/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neocortex-livekit"
version = "0.1.0"
description = "LiveKit plugin for Neocortex-powered memories in LiveKit agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-silero>=1.4.0",
    "livekit-plugins-turn-detector>=1.4.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.setuptools.packages.find]
include = ["neocortex_livekit*"]

[tool.ruff]
line-length = 88
target-version = "py39"
